requirements.txt
psutil
asciimatics
GPUtil
python-chess==0.31.4
matplotlib
pytorch-lightning==1.9.5
tensorboard
cupy-cuda11x
numba
